+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact limestone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `limestone/ImageItem` to get custom `aria-label`
+
 ## [1.6.0] - 2025-10-14
 
 ### Added